+++ py/desisurvey/test/test_schedule.py
@@ -7,7 +7,7 @@ import numpy as np
 
 import desisurvey.config
 
-from desisurvey.plan import *
+from desisurvey.schedule import *
 
 
 class TestPlan(unittest.TestCase):
@@ -28,7 +28,7 @@ class TestPlan(unittest.TestCase):
         desisurvey.config.Configuration.reset()
 
     def test_initialize(self):
-        """Create a temporary planner (requires healpy)"""
+        """Create a temporary scheduler (requires healpy)"""
         start = desisurvey.utils.get_date('2020-01-01')
         stop = desisurvey.utils.get_date('2020-01-03')
         ephem = desisurvey.ephemerides.Ephemerides(start, stop)
@@ -36,21 +36,21 @@ class TestPlan(unittest.TestCase):
 
     def test_lst_range(self):
         """Test that 0 <= LST < 360 deg"""
-        p = Planner(os.path.join(self.tmpdir, 'planner.fits'))
+        p = Scheduler(os.path.join(self.tmpdir, 'scheduler.fits'))
         lst = p.etable['lst']
         self.assertTrue(np.min(lst) >= 0.)
         self.assertTrue(np.max(lst) < 360.)
 
     def test_time_index_conversion(self):
         """Test time <-> index round trips"""
-        p = Planner(os.path.join(self.tmpdir, 'planner.fits'))
+        p = Scheduler(os.path.join(self.tmpdir, 'scheduler.fits'))
         for i in range(p.num_nights * p.num_times):
             t = p.time_of_index(i)
             self.assertEqual(i, p.index_of_time(t))
 
     def test_spatial_index_conversion(self):
         """Test tile_id -> index"""
-        p = Planner(os.path.join(self.tmpdir, 'planner.fits'))
+        p = Scheduler(os.path.join(self.tmpdir, 'scheduler.fits'))
         for i in range(len(p.tiles)):
             tid = p.tiles['tileid'][i]
             pix = p.tiles['map'][i]