+++ doc/changes.rst
@@ -5,14 +5,16 @@ desisurvey change log
 0.7.1 (unreleased)
 ------------------
 
-* Implement iterative HA optimization and LST-driven scheduler.
+* Implement LST-driven scheduling strategy.
+* Create new optimize module for iterative HA optimization.
+* Rename module plan -> schedule.
 
 0.7.0 (2017-06-05)
 ------------------
 
 * Freeze IERS table used by astropy time, coordinates.
 * Implement alternate greedy scheduler with optional policy weights.
-* Add `plots.plot_planner()`
+* Add `plots.plot_scheduler()`
 * Partial fix of RA=0/360 planning bug
 
 0.6.0 (2017-05-10)